Exercises 4.21 and 4.22

+++ osa4/blogilista/controllers/blogs.js
@@ -3,6 +3,7 @@ const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
 const User = require('../models/user')
 const logger = require('../utils/logger')
+const { userExtractor } = require('../utils/middleware')
 
 blogsRouter.get('/', async (request, response) => {
   const blogs = await Blog.find({})
@@ -21,22 +22,16 @@ blogsRouter.get('/:id', async (request, response) => {
   response.json(blog)
 })
 
-blogsRouter.post('/', async (request, response) => {
+blogsRouter.post('/', userExtractor, async (request, response) => {
 
   const { title, author, url, likes } = request.body
-
-  const decodedToken = jwt.verify(request.token, process.env.SECRET)
-  if (!decodedToken.id) {
-    return response.status(401).json({ error: 'token invalid' })
-  }
-  const user = await User.findById(decodedToken.id)
 
   const blog = new Blog({
     title: title,
     author: author,
     url: url,
     likes: likes,
-    user: user._id
+    user: request.user._id
   })
 
   if(!blog.likes) blog.likes=0
@@ -49,8 +44,8 @@ blogsRouter.post('/', async (request, response) => {
 
   else {
     const savedBlog = await blog.save()
-    user.blogs = user.blogs.concat(savedBlog._id)
-    await user.save()
+    request.user.blogs = request.user.blogs.concat(savedBlog._id)
+    await request.user.save()
     response.status(201).json(savedBlog)
   }
 })
@@ -71,13 +66,20 @@ blogsRouter.put('/:id', async (request, response) => {
 })
 
 
-blogsRouter.delete('/:id', async (request, response) => {
-
+blogsRouter.delete('/:id', userExtractor, async (request, response) => {
+  
   const blog = await Blog.findById(request.params.id)
   
   if (!blog) {
     return response.status(404).end()  
   }
+
+  if (request.user.id.toString() != blog.user.toString())
+  {
+    return response.status(403).json({ error: 'only user, who had added, can remove' })
+  }
+
+  console.log('deleting..')
 
   await Blog.findByIdAndDelete(request.params.id)
   response.status(204).end()
